Revert some changes that might break compat but don't seem to break the game

+++ src/d_clisrv.h
@@ -423,7 +423,7 @@ typedef struct
 	char name[MAXPLAYERNAME+1];
 	UINT8 address[4]; // sending another string would run us up against MAXPACKETLENGTH
 	UINT8 team;
-	INT32 skin;
+	UINT8 skin;
 	UINT8 data; // Color is first four bits, hasflag, isit and issuper have one bit each, the last is unused.
 	UINT32 score;
 	UINT16 timeinserver; // In seconds.
@@ -433,7 +433,7 @@ typedef struct
 typedef struct
 {
 	char name[MAXPLAYERNAME+1];
-	INT32 skin;
+	UINT8 skin;
 	UINT8 color;
 	UINT32 pflags;
 	UINT32 score;
